app/applipsync/serializers.py
from rest_framework import serializers

from .models import File, VideoFrame


class FileSerializer(serializers.ModelSerializer):
    class Meta:
        model = File
        fields = "__all__"


class VideoFrameSerializer(serializers.ModelSerializer):
    class Meta:
        model = VideoFrame
        fields = "__all__"
