app/requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
autopep8==2.0.0
blessed==1.19.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
defusedxml==0.7.1
Distance==0.1.3
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
g2p-en==2.1.0
idna==3.4
inflect==6.0.2
install==1.3.5
jinxed==1.2.0
joblib==1.2.0
jsonfield==3.1.0
Markdown==3.4.1
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==3.5.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
wcwidth==0.2.5
psycopg2-binary==2.9.1
gunicorn==20.1.0
black==22.12.0
flake8==6.0.0
isort==5.11.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.10.0
pyflakes==3.0.1